refactor(tests): Reuse waitAndFetchJobResult
The method has been used in slight variation across multiple tests.
Unify it into a common function in JobSpecBase and reuse this
central method in every test.
