+++ job-server-extras/src/test/scala/spark/jobserver/NamedObjectsJobSpec.scala
@@ -1,9 +1,7 @@
 package spark.jobserver
 
-import com.typesafe.config.{Config, ConfigFactory, ConfigValueFactory}
+import com.typesafe.config.{Config, ConfigFactory}
 import akka.testkit.TestProbe
-import spark.jobserver.CommonMessages.{JobFinished, JobStarted}
-import spark.jobserver.io.JobDAOActor.{GetJobResult, JobResult}
 import spark.jobserver.io.{BinaryInfo, InMemoryBinaryObjectsDAO, InMemoryMetaDAO, JobDAOActor}
 
 class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) {
@@ -12,7 +10,6 @@ class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) 
   lazy val cfg = JobManagerActorSpec.getContextConfig(adhoc = false)
 
   var testBinInfo: BinaryInfo = _
-  val smallTimeout = 5.seconds
 
   override def beforeAll() {
     inMemoryMetaDAO = new InMemoryMetaDAO
@@ -42,38 +39,27 @@ class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) 
         NamedObjectsTestJobConfig.DELETE+" = [" + names.mkString(", ") + "]")
   }
 
-  private def waitAndFetchJobResult(): Array[String] = {
-    expectMsgPF(smallTimeout, "Never got a JobStarted event") {
-      case JobStarted(jobId, result) =>
-        expectMsgClass(classOf[JobFinished])
-        daoActor ! GetJobResult(jobId)
-        expectMsgPF(smallTimeout, "Never got a JobResult") {
-          case JobResult(result) => result match {
-            case result: Array[String] => result
-            case _ => throw new Exception("Unexpected result type!")
-          }
-        }
-      case message: Any => throw new Exception(s"Got unexpected message $message")
-    }
+  private def waitAndFetchJobResultArray(jobFinishTimeout: FiniteDuration = 10 seconds): Array[String] = {
+    super.waitAndFetchJobResult(jobFinishTimeout).asInstanceOf[Array[String]]
   }
 
   describe("NamedObjects (RDD)") {
     it("should survive from one job to another one") {
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(false, true), allEvents)
-      val firstJobResult = waitAndFetchJobResult()
+      val firstJobResult = waitAndFetchJobResultArray()
       firstJobResult should contain("rdd1")
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(false, false), allEvents)
-      val secondJobResult = waitAndFetchJobResult()
+      val secondJobResult = waitAndFetchJobResultArray()
       secondJobResult should contain("rdd1")
       secondJobResult should not contain("df1")
 
       //clean-up
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getDeleteConfig(List("rdd1")), allEvents)
-      val thirdJobResult = waitAndFetchJobResult()
+      val thirdJobResult = waitAndFetchJobResultArray()
       thirdJobResult should not contain("rdd1")
       thirdJobResult should not contain("df1")
     }
@@ -84,18 +70,18 @@ class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) 
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(true, false), allEvents)
-      val firstJobResult = waitAndFetchJobResult()
+      val firstJobResult = waitAndFetchJobResultArray()
       firstJobResult should contain("df1")
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(false, false), allEvents)
-      val secondJobResult = waitAndFetchJobResult()
+      val secondJobResult = waitAndFetchJobResultArray()
       secondJobResult should equal(firstJobResult)
 
       //clean-up
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getDeleteConfig(List("df1")), allEvents)
-      waitAndFetchJobResult()
+      waitAndFetchJobResultArray()
     }
   }
 
@@ -105,19 +91,19 @@ class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) 
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(true, true), allEvents)
-      val firstJobResult = waitAndFetchJobResult()
+      val firstJobResult = waitAndFetchJobResultArray()
       firstJobResult should contain("rdd1")
       firstJobResult should contain("df1")
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(false, false), allEvents)
-      val secondJobResult = waitAndFetchJobResult()
+      val secondJobResult = waitAndFetchJobResultArray()
       secondJobResult should equal(firstJobResult)
 
       //clean-up
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getDeleteConfig(List("rdd1", "df1")), allEvents)
-      waitAndFetchJobResult()
+      waitAndFetchJobResultArray()
     }
   }
 
@@ -127,21 +113,21 @@ class NamedObjectsJobSpec extends JobSpecBase(JobManagerActorSpec.getNewSystem) 
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(true, true, true), allEvents)
-      val firstJobResult = waitAndFetchJobResult()
+      val firstJobResult = waitAndFetchJobResultArray()
       firstJobResult should contain("rdd1")
       firstJobResult should contain("df1")
       firstJobResult should contain("broadcast1")
 
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getCreateConfig(false, false, false), allEvents)
-      val secondJobResult = waitAndFetchJobResult()
+      val secondJobResult = waitAndFetchJobResultArray()
       secondJobResult should equal(firstJobResult)
 
       //clean-up
       manager ! JobManagerActor.StartJob(
         jobName, Seq(testBinInfo), getDeleteConfig(List("rdd1", "df1", "broadcast1")),
         allEvents)
-      waitAndFetchJobResult()
+      waitAndFetchJobResultArray()
     }
   }
 }
